+++ test/test_helper.rb
@@ -5,7 +5,8 @@ require File.expand_path(File.dirname(__FILE__) + '/../../../test/test_helper')
 require 'minitest/spec'
 require 'minitest/mock'
 require 'minitest/autorun'
-require "minitest/reporters"
+require 'minitest/reporters'
+require 'telegram/bot'
 
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new